Add reset button to filter modal
