+++ src/actions/index.js
@@ -5,6 +5,7 @@ export const GET_TOP_RATED = 'GET_TOP_RATED';
 export const ADD_FILTER = 'ADD_FILTER';
 export const GET_VIDEOS = 'GET_VIDEOS';
 export const REMOVE_FILTER = 'REMOVE_FILTER';
+export const RESET_FILTER = 'RESET_FILTER';
 
 export const loadTrending = movie => ({ type: GET_TRENDING, movie });
 export const loadVideos = trailer => ({ type: GET_VIDEOS, trailer });
@@ -13,3 +14,4 @@ export const loadTopRated = movies => ({ type: GET_TOP_RATED, movies });
 export const loadUpcoming = movies => ({ type: GET_UPCOMING, movies });
 export const addFilter = filter => ({ type: ADD_FILTER, filter });
 export const removeFilter = filter => ({ type: REMOVE_FILTER, filter });
+export const resetFilter = filter => ({ type: RESET_FILTER, filter });
