+++ src/containers/FilterModal.js
@@ -1,15 +1,32 @@
 import React, { useState } from 'react';
 import Modal from 'react-modal';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faFilter, faTimes } from '@fortawesome/free-solid-svg-icons';
+import { addFilter, removeFilter, resetFilter } from '../actions';
 import '../styles/filtermodal/FilterModal.css';
 import FilterOptions from '../components/FilterOptions';
 
-const FilterModal = () => {
+const FilterModal = ({
+  filters, addFilters, removeFilters, resetFilters,
+}) => {
   const [showModal, setShowModal] = useState(false);
 
   const switcher = () => {
     setShowModal(!showModal);
+  };
+
+  const handleChange = e => {
+    const { id, checked } = e.target;
+    return checked ? addFilters(id) : removeFilters(id);
+  };
+
+  const isChecked = id => !!filters.includes(id);
+
+  const resetFilterHandler = () => {
+    resetFilters();
+    switcher();
   };
 
   return (
@@ -23,10 +40,26 @@ const FilterModal = () => {
           <h1>Filter</h1>
         </div>
 
-        <FilterOptions />
+        <FilterOptions handleChange={handleChange} isChecked={isChecked} />
+        <button onClick={resetFilterHandler} type="button">RESET</button>
       </Modal>
     </div>
   );
 };
 
-export default FilterModal;
+FilterModal.propTypes = {
+  filters: PropTypes.oneOfType([PropTypes.object, PropTypes.array]).isRequired,
+  addFilters: PropTypes.func.isRequired,
+  removeFilters: PropTypes.func.isRequired,
+  resetFilters: PropTypes.func.isRequired,
+};
+
+const mapStateToProps = state => ({ filters: state.filter });
+
+const mapDispatchToProps = dispatch => ({
+  addFilters: filter => dispatch(addFilter(filter)),
+  removeFilters: filter => dispatch(removeFilter(filter)),
+  resetFilters: filter => dispatch(resetFilter(filter)),
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(FilterModal);
